refactor, add fonts, details
