+++ src/components/landing/rework/menu-new.tsx
@@ -4,10 +4,10 @@ import type { TippyProps } from '@tippyjs/react'
 import { useState, useEffect } from 'react'
 import styled from 'styled-components'
 
-import { Link } from '../content/link'
-import { SubButtonStyle } from '../user-tools/sub-button-style'
-import { getAvatarUrl } from '../user/user-link'
-import { SubLink } from './sub-link'
+import { Link } from '../../content/link'
+import { SubLink } from '../../navigation/sub-link'
+import { SubButtonStyle } from '../../user-tools/sub-button-style'
+import { getAvatarUrl } from '../../user/user-link'
 import { AuthenticationPayload } from '@/auth/use-authentication'
 import { useInstanceData } from '@/contexts/instance-context'
 import { useLoggedInComponents } from '@/contexts/logged-in-components'
