+++ src/components/landing/rework/partner-list-new.tsx
