+++ src/components/landing/rework/separator.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const Separator = styled.hr`
+  border: 0;
+  outline: none;
+  background: url('/_assets/img/landing/separator.svg') no-repeat center center;
+  height: 2rem;
+  width: 80%;
+  margin: 0 auto;
+  padding: 5rem 0;
+`
